api/publicAPI/requirements.txt
fastapi
uvicorn
requests
python-dotenv
cachetools